requirements.txt
supabase==1.2.0
pydantic==1.10.20
cryptography
python-dotenv
python-telegram-bot==13.15
urllib3